change script naming convention
